Update theme (#17)
* inital work

* Expose ThemeContext publicly

* Removed useless complexity

* More clean-up work

* Simplify withTheme.lua's context implementation.
Need not do it all in the Provider. The consumer itself can do it.
Instead of checking if we have a theme in the Provider (in a hacky way).
We can check it when trying to consume. If nil, then there was no
provider.
Which case, we can go ahead and make ours.